requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2
APScheduler==3.10.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.2
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitarray==3.0.0
black==24.8.0
blinker==1.8.2
boto3==1.35.0
botocore==1.35.54
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.9
click==8.1.7
colbert-ai==0.2.21
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
cryptography==43.0.3
cssselect2==0.7.0
ctranslate2==4.5.0
dataclasses-json==0.6.7
datasets==3.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckduckgo_search==6.2.13
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
emoji==2.14.0
environs==9.5.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
evaluate==0.4.3
extract-msg==0.52.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.5
faster-whisper==1.0.3
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf2==2.7.9
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.2.15
langchain-chroma==0.1.4
langchain-community==0.2.12
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.1.139
lark==1.1.9
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
milvus-lite==2.4.10
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.20.0
open-webui==0.3.35
openai==1.54.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.11
oscrypto==1.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
peewee==3.17.6
peewee-migrate==1.12.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.7.0
primp==0.7.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pymdown-extensions==10.11.2
pymilvus==2.4.7
pymongo==4.10.1
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.2.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-docker==3.1.1
python-bidi==0.6.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-iso639==2024.10.22
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.0
python-socketio==5.11.3
pytube==15.0.0
pytz==2024.2
pyxlsb==1.0.10
PyYAML==6.0.2
qrcode==8.0
rank-bm25==0.2.2
RapidFuzz==3.10.1
rapidocr-onnxruntime==1.3.24
red-black-tree-mod==1.20
redis==5.2.0
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
RTFDE==0.1.2
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
starlette==0.37.2
svglib==1.5.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
unstructured==0.15.9
unstructured-client==0.27.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
validators==0.33.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.2
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.16
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.15.5
youtube-transcript-api==0.6.2
zipp==3.20.2
